+++ gateway/src/main/java/com/mustrysolutions/perspective/components/gateway/MustrySolutionsPerspectiveComponentsGatewayHook.java
@@ -13,6 +13,7 @@ import com.inductiveautomation.perspective.gateway.api.PerspectiveContext;
 
 import com.mustrysolutions.perspective.components.common.comp.Calendar;
 import com.mustrysolutions.perspective.components.common.comp.DataGrid;
+import com.mustrysolutions.perspective.components.common.comp.PanZoomView;
 import com.mustrysolutions.perspective.components.common.comp.DateTimeRangePicker;
 import com.mustrysolutions.perspective.components.common.comp.ResourceTimeline;
 
@@ -44,6 +45,7 @@ public class MustrySolutionsPerspectiveComponentsGatewayHook extends AbstractGat
             this.componentRegistry.registerComponent(DateTimeRangePicker.DESCRIPTOR);
             this.componentRegistry.registerComponent(Calendar.DESCRIPTOR);
             this.componentRegistry.registerComponent(DataGrid.DESCRIPTOR);
+            this.componentRegistry.registerComponent(PanZoomView.DESCRIPTOR);
             this.componentRegistry.registerComponent(ResourceTimeline.DESCRIPTOR);
         } else {
             log.error("Perspective component registry not found; components not registered.");
@@ -56,6 +58,7 @@ public class MustrySolutionsPerspectiveComponentsGatewayHook extends AbstractGat
             this.componentRegistry.removeComponent(DateTimeRangePicker.COMPONENT_ID);
             this.componentRegistry.removeComponent(Calendar.COMPONENT_ID);
             this.componentRegistry.removeComponent(DataGrid.COMPONENT_ID);
+            this.componentRegistry.removeComponent(PanZoomView.COMPONENT_ID);
             this.componentRegistry.removeComponent(ResourceTimeline.COMPONENT_ID);
         }
     }